fixed and consolidated userTypeEnum
