--- src/modules/authentication/enums/user.enum.ts
@@ -1,5 +1,0 @@
-export enum UserEnum {
-  customer = 'customer',
-  seller = 'seller',
-  admin = 'admin',
-}
