+++ src/modules/authorization/dtos/create-role.dto.ts
@@ -6,7 +6,7 @@ import {
   IsBoolean,
 } from 'class-validator';
 import { RoleInterface } from '../interfaces/role.interface';
-import { UserTypeEnum } from '../enums/user-type.enum';
+import { UserTypeEnum } from '../../../utils/enums/user-type.enum';
 
 export class CreateRoleDto implements RoleInterface {
   @IsNotEmpty()
