+++ src/utils/dtos/device-token.dto.ts
@@ -2,7 +2,7 @@ import { IsString, IsNotEmpty, IsBoolean, IsNumber } from 'class-validator';
 import { ApplicationPlatformEnum } from '../enums/application-platform.enum';
 
 import { DeviceTokenInterface } from '../interfaces/device-token.interface';
-import { UserTypeEnum } from '@modules/authorization/enums/user-type.enum';
+import { UserTypeEnum } from '@src/utils/enums/user-type.enum';
 
 export class DeviceTokenDto implements DeviceTokenInterface{
   @IsString()
