+++ src/utils/interfaces/event.interface.ts
@@ -1,4 +1,4 @@
-import { UserTypeEnum } from '@modules/authorization/enums/user-type.enum';
+import { UserTypeEnum } from '@src/utils/enums/user-type.enum';
 
 export interface EventInterface {
   id: string;
