+++ src/utils/notification/interfaces/event.configuration.interface.ts
@@ -1,5 +1,5 @@
 import { NotificationTypeEnum } from './../enums/notification-type.enum';
-import { UserTypeEnum } from '@modules/authorization/enums/user-type.enum';
+import { UserTypeEnum } from '@src/utils/enums/user-type.enum';
 import { EventNameEnum } from '../enums/event-name.enum';
 import { EmailTemplateInterface } from './email-template.interface';
