+++ src/utils/services/push-notification-builder.service.ts
@@ -5,7 +5,7 @@ import { EventConfigurationInterface } from '../interfaces/event-configuration.i
 import * as Handlebars from 'handlebars';
 import { PushNotificationService } from './push-notification.service';
 import { DeviceTokenRepository } from '../repositories/device-token.repository';
-import { UserTypeEnum } from '@modules/authorization/enums/user-type.enum';
+import { UserTypeEnum } from '@src/utils/enums/user-type.enum';
 
 
 @Injectable()
